requirements.txt
Flask==2.0.0
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.25.0
wikipedia==1.4.0
sentence-transformers==2.2.0
PyPDF2==3.0.0
spacy==3.8.0